==> notebooks/06-prepare-upcoming-games-data.ipynb <==
import pandas as pd
import numpy as np
import pickle
pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 500)
pd.set_option('display.width', 150)

df = pd.read_pickle('/Users/andrewpeters/GitHub/fpl/data/interim/df_with_rolling_features.pkl')
unplayed = pd.read_pickle('/Users/andrewpeters/GitHub/fpl/data/interim/unplayed.pkl')
team_level = pd.read_pickle('/Users/andrewpeters/GitHub/fpl/data/interim/team_level_rolling.pkl')

recent_avg_minutes = df.sort_values('kickoff_time').groupby('player').tail(3).groupby('player')['minutes'].mean().reset_index()

players_not_playing = recent_avg_minutes[recent_avg_minutes['minutes'] < 60]['player']

last_performance = df.sort_values(['kickoff_time', 'player']).drop_duplicates(subset='player', keep='last')

#only want to keep cols that have data about the player/teams's previous performance
#drop cols that are related to the opponent or the specific game at hand

keep_cols = ['player',
     'total_points_prev',
     'team_h_score_prev',
     'team_a_score_prev',
     'minutes_prev',
     'goals_scored_prev',
     'assists_prev',
     'clean_sheets_prev',
     'goals_conceded_prev',
     'bonus_prev',
     'bps_prev',
     'influence_prev',
     'creativity_prev',
     'threat_prev',
     'ict_index_prev',
     'transfers_balance_prev',
     'selected_prev',
     'transfers_in_prev',
     'transfers_out_prev',
     'total_points_prev_2',
     'team_h_score_prev_2',
     'team_a_score_prev_2',
     'minutes_prev_2',
     'goals_scored_prev_2',
     'assists_prev_2',
     'clean_sheets_prev_2',
     'goals_conceded_prev_2',
     'bonus_prev_2',
     'bps_prev_2',
     'influence_prev_2',
     'creativity_prev_2',
     'threat_prev_2',
     'ict_index_prev_2',
     'transfers_balance_prev_2',
     'selected_prev_2',
     'transfers_in_prev_2',
     'transfers_out_prev_2',
     'total_points_prev_3',
     'team_h_score_prev_3',
     'team_a_score_prev_3',
     'minutes_prev_3',
     'goals_scored_prev_3',
     'assists_prev_3',
     'clean_sheets_prev_3',
     'goals_conceded_prev_3',
     'bonus_prev_3',
     'bps_prev_3',
     'influence_prev_3',
     'creativity_prev_3',
     'threat_prev_3',
     'ict_index_prev_3',
     'transfers_balance_prev_3',
     'selected_prev_3',
     'transfers_in_prev_3',
     'transfers_out_prev_3',
     'total_points_prev_4',
     'team_h_score_prev_4',
     'team_a_score_prev_4',
     'minutes_prev_4',
     'goals_scored_prev_4',
     'assists_prev_4',
     'clean_sheets_prev_4',
     'goals_conceded_prev_4',
     'bonus_prev_4',
     'bps_prev_4',
     'influence_prev_4',
     'creativity_prev_4',
     'threat_prev_4',
     'ict_index_prev_4',
     'transfers_balance_prev_4',
     'selected_prev_4',
     'transfers_in_prev_4',
     'transfers_out_prev_4',
     'total_points_prev_5',
     'team_h_score_prev_5',
     'team_a_score_prev_5',
     'minutes_prev_5',
     'goals_scored_prev_5',
     'assists_prev_5',
     'clean_sheets_prev_5',
     'goals_conceded_prev_5',
     'bonus_prev_5',
     'bps_prev_5',
     'influence_prev_5',
     'creativity_prev_5',
     'threat_prev_5',
     'ict_index_prev_5',
     'transfers_balance_prev_5',
     'selected_prev_5',
     'transfers_in_prev_5',
     'transfers_out_prev_5',
     'total_points_prev_6',
     'team_h_score_prev_6',
     'team_a_score_prev_6',
     'minutes_prev_6',
     'goals_scored_prev_6',
     'assists_prev_6',
     'clean_sheets_prev_6',
     'goals_conceded_prev_6',
     'bonus_prev_6',
     'bps_prev_6',
     'influence_prev_6',
     'creativity_prev_6',
     'threat_prev_6',
     'ict_index_prev_6',
     'transfers_balance_prev_6',
     'selected_prev_6',
     'transfers_in_prev_6',
     'transfers_out_prev_6',
     'total_points_prev_7',
     'team_h_score_prev_7',
     'team_a_score_prev_7',
     'minutes_prev_7',
     'goals_scored_prev_7',
     'assists_prev_7',
     'clean_sheets_prev_7',
     'goals_conceded_prev_7',
     'bonus_prev_7',
     'bps_prev_7',
     'influence_prev_7',
     'creativity_prev_7',
     'threat_prev_7',
     'ict_index_prev_7',
     'transfers_balance_prev_7',
     'selected_prev_7',
     'transfers_in_prev_7',
     'transfers_out_prev_7',
     'total_points_prev_8',
     'team_h_score_prev_8',
     'team_a_score_prev_8',
     'minutes_prev_8',
     'goals_scored_prev_8',
     'assists_prev_8',
     'clean_sheets_prev_8',
     'goals_conceded_prev_8',
     'bonus_prev_8',
     'bps_prev_8',
     'influence_prev_8',
     'creativity_prev_8',
     'threat_prev_8',
     'ict_index_prev_8',
     'transfers_balance_prev_8',
     'selected_prev_8',
     'transfers_in_prev_8',
     'transfers_out_prev_8',
     'total_points_prev_9',
     'team_h_score_prev_9',
     'team_a_score_prev_9',
     'minutes_prev_9',
     'goals_scored_prev_9',
     'assists_prev_9',
     'clean_sheets_prev_9',
     'goals_conceded_prev_9',
     'bonus_prev_9',
     'bps_prev_9',
     'influence_prev_9',
     'creativity_prev_9',
     'threat_prev_9',
     'ict_index_prev_9',
     'transfers_balance_prev_9',
     'selected_prev_9',
     'transfers_in_prev_9',
     'transfers_out_prev_9',
     'total_points_prev_10',
     'team_h_score_prev_10',
     'team_a_score_prev_10',
     'minutes_prev_10',
     'goals_scored_prev_10',
     'assists_prev_10',
     'clean_sheets_prev_10',
     'goals_conceded_prev_10',
     'bonus_prev_10',
     'bps_prev_10',
     'influence_prev_10',
     'creativity_prev_10',
     'threat_prev_10',
     'ict_index_prev_10',
     'transfers_balance_prev_10',
     'selected_prev_10',
     'transfers_in_prev_10',
     'transfers_out_prev_10',
     'total_points_team',
     'total_points_prev_team',
     'goals_scored_prev_team',
     'clean_sheets_prev_team',
     'goals_conceded_prev_team',
     'total_points_prev_2_team',
     'goals_scored_prev_2_team',
     'clean_sheets_prev_2_team',
     'goals_conceded_prev_2_team',
     'influence_prev_2_team',
     'creativity_prev_2_team',
     'threat_prev_2_team',
     'total_points_prev_3_team',
     'goals_scored_prev_3_team',
     'clean_sheets_prev_3_team',
     'goals_conceded_prev_3_team',
     'influence_prev_3_team',
     'creativity_prev_3_team',
     'threat_prev_3_team',
     'total_points_prev_4_team',
     'goals_scored_prev_4_team',
     'clean_sheets_prev_4_team',
     'goals_conceded_prev_4_team',
     'influence_prev_4_team',
     'creativity_prev_4_team',
     'threat_prev_4_team',
     'total_points_prev_5_team',
     'goals_scored_prev_5_team',
     'clean_sheets_prev_5_team',
     'goals_conceded_prev_5_team',
     'influence_prev_5_team',
     'creativity_prev_5_team',
     'threat_prev_5_team',
     'total_points_prev_6_team',
     'goals_scored_prev_6_team',
     'clean_sheets_prev_6_team',
     'goals_conceded_prev_6_team',
     'influence_prev_6_team',
     'creativity_prev_6_team',
     'threat_prev_6_team',
     'total_points_prev_7_team',
     'goals_scored_prev_7_team',
     'clean_sheets_prev_7_team',
     'goals_conceded_prev_7_team',
     'influence_prev_7_team',
     'creativity_prev_7_team',
     'threat_prev_7_team',
     'total_points_prev_8_team',
     'goals_scored_prev_8_team',
     'clean_sheets_prev_8_team',
     'goals_conceded_prev_8_team',
     'influence_prev_8_team',
     'creativity_prev_8_team',
     'threat_prev_8_team',
     'total_points_prev_9_team',
     'goals_scored_prev_9_team',
     'clean_sheets_prev_9_team',
     'goals_conceded_prev_9_team',
     'influence_prev_9_team',
     'creativity_prev_9_team',
     'threat_prev_9_team',
     'total_points_prev_10_team',
     'goals_scored_prev_10_team',
     'clean_sheets_prev_10_team',
     'goals_conceded_prev_10_team',
     'influence_prev_10_team',
     'creativity_prev_10_team',
     'threat_prev_10_team']

#drop players who are in our 'not_playing_list' -- this is consistent with our training workflow, which drops players who haven't averaged 60 mins over the past 3 games
unplayed = unplayed[~unplayed.player.isin(players_not_playing)]

unplayed = unplayed.merge(last_performance[keep_cols], how='left', on='player')

last_team_level = team_level.sort_values(['team', 'season', 'gw']).drop_duplicates(subset='team', keep='last')

unplayed = unplayed.merge(last_team_level, how='left', left_on = 'opponent_team', right_on='team', suffixes = ('', '_opponent'))

All the following work is done to make the features and feature names match what was used to train the model

unplayed = unplayed.rename(columns={'is_home': 'was_home'})

unplayed['was_home'] = unplayed['was_home'].astype('int') #convert bool to numeric

positions_df = pd.get_dummies(unplayed['position'], prefix='position_')
unplayed = unplayed.join(positions_df)

len(unplayed)

unplayed = unplayed.dropna()

len(unplayed)

metadata_to_keep = ['element',
                   'kickoff_time',
                   'gw',
                   'player',
                   'position',
                   'team',
                   'opponent_team']

#need to keep this info so that once we predict points later, we know which players those predictions belong to
unplayed_metadata = unplayed.loc[:, metadata_to_keep]
unplayed_metadata.to_pickle('/Users/andrewpeters/GitHub/fpl/data/processed/unplayed_metadata.pkl')

unplayed = unplayed.drop(columns = metadata_to_keep 
                         + ['finished',
                             'team_opponent',
                             'season',
                             'gw_opponent',
                             'total_points'])

with open('/Users/andrewpeters/GitHub/fpl/data/interim/x_features.pkl', 'rb') as f:
    feature_list = pickle.load(f)

#verify that the unplayed columns match up with the features used to train the model
list(unplayed) == feature_list

from sklearn.preprocessing import MinMaxScaler
scaler = MinMaxScaler()
unplayed = scaler.fit_transform(unplayed)

with open('/Users/andrewpeters/GitHub/fpl/data/processed/unplayed.npy', 'wb') as file:
    np.save(file, unplayed)
