==> notebooks/07-predict-unplayed-games.ipynb <==
import pandas as pd
import numpy as np
import pickle
import matplotlib.pyplot as plt
from sklearn.preprocessing import MinMaxScaler
pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 500)
pd.set_option('display.width', 150)

with open('/Users/andrewpeters/GitHub/fpl/data/processed/unplayed.npy', 'rb') as file:
    X = np.load(file)

with open('/Users/andrewpeters/GitHub/fpl/models/xgb_256.pkl', 'rb') as file:
    model = pickle.load(file)

with open ('/Users/andrewpeters/GitHub/fpl/models/min_max_scaler.pkl', 'rb') as file:
    scaler = pickle.load(file)

X = scaler.transform(X)

predicted_points = model.predict(X)

metadata = pd.read_pickle('/Users/andrewpeters/GitHub/fpl/data/processed/unplayed_metadata.pkl')

metadata['predicted_points'] = predicted_points

metadata.to_pickle('/Users/andrewpeters/GitHub/fpl/data/processed/predictions.pkl')
