==> notebooks/09-optimize-subs.ipynb <==
import requests
import pandas as pd
import numpy as np
import pulp

# Use API to Grab Current Team

history = pd.read_pickle('/Users/andrewpeters/GitHub/fpl/data/interim/df.pkl')
current_gw = int(history[history['finished'] == True].sort_values('kickoff_time')['gw'].tail(1))

url = f'https://fantasy.premierleague.com/api/entry/5171826/event/{current_gw}/picks/'
r = requests.get(url)
json = r.json()

current_team = pd.DataFrame(json['picks'])
money_remaining = json['entry_history']['bank']

current_team = current_team.merge(history[['element', 'player']], how='left').drop_duplicates().reset_index(drop=True)

# Solve Knapsack Problem

df = pd.read_pickle('/Users/andrewpeters/GitHub/fpl/data/processed/grouped_predictions_with_cost.pkl')

#currently not calculating for change in player value

df.gw = df.gw.astype('int64')
data = df[df.gw.isin(df.gw.unique()[:5])] #filter for next 5 gws only
data = data.groupby(['player', 'position', 'team', 'now_cost', 'element'])['predicted_points'].sum().reset_index()

expected_scores = data.predicted_points.values
num_players = len(data)
buy_prices = data.now_cost.values / 10
sell_prices = data.now_cost.values / 10
positions = data.position.values
teams = data.team.values
names = data.player.values
budget_now = money_remaining/10

from transfer_optimizer import TransferOptimiser

opt = TransferOptimiser(expected_scores, buy_prices, sell_prices, positions, teams)

current_squad_indices = data[data.element.isin(current_team.element)].index #look up indices for current team

transfer_in_decisions, transfer_out_decisions, starters, sub_decisions, captain_decisions = opt.solve(current_squad_indices, budget_now, sub_factor=0.2)

for i in range(num_players):
    if transfer_in_decisions[i].value() == 1:
        print("Transferred in: {} {} {:.2f}".format(names[i], buy_prices[i], expected_scores[i]))
    if transfer_out_decisions[i].value() == 1:
        print("Transferred out: {} {} {:.2f}".format(names[i], sell_prices[i], expected_scores[i]))

player_indices = []
print()
print("First Team:")
for i in range(len(starters)):
    if starters[i].value() == 1:
        print("{}{}".format(names[i], "*" if captain_decisions[i].value() == 1 else ""))
        player_indices.append(i)
print()
print("Subs:")
for i in range(len(sub_decisions)):
    if sub_decisions[i].value() == 1:
        print(names[i])
        player_indices.append(i)
